Moved top tracks into seperate file, working but crashing immediately after
