+++ backend/server.js
@@ -18,6 +18,14 @@ const spotifyApi = new SpotifyWebApi({
   redirectUri: process.env.REDIRECT_URI,
 });
 
+// Import routes
+//const detailedStatsRoute = require('./routes/detailedStats')(spotifyApi);
+import topTracksRoute from './routes/topTracks.js';
+
+// Use the routes
+//app.use(detailedStatsRoute);
+app.use(topTracksRoute(spotifyApi));
+
 // route for login authentication
 app.get("/login", (req, res) => {
   const scopes = [
@@ -92,40 +100,40 @@ app.get("/search", async (req, res) => {
     });
 });
 
-//top-tracks
-app.get("/top-tracks", async (req, res) => {
-  const { term } = req.query;
-  const validTerms = ["short_term", "medium_term", "long_term"];
-  const timeRange = validTerms.includes(term) ? term : "medium_term";
+// //top-tracks
+// app.get("/top-tracks", async (req, res) => {
+//   const { term } = req.query;
+//   const validTerms = ["short_term", "medium_term", "long_term"];
+//   const timeRange = validTerms.includes(term) ? term : "medium_term";
 
-  try {
-    const topTracksResponse = await spotifyApi.getMyTopTracks({
-      time_range: timeRange,
-    });
-    const topTracks = topTracksResponse.body.items;
-    const top10Tracks = topTracks.slice(0, 10).map((track) => ({
-      name: track.name,
-      album: track.album.name,
-      artists: track.artists.map((artist) => artist.name),
-      popularity: track.popularity,
-      externalUrl: track.external_urls.spotify,
-      images: track.album.images,
-    }));
+//   try {
+//     const topTracksResponse = await spotifyApi.getMyTopTracks({
+//       time_range: timeRange,
+//     });
+//     const topTracks = topTracksResponse.body.items;
+//     const top10Tracks = topTracks.slice(0, 10).map((track) => ({
+//       name: track.name,
+//       album: track.album.name,
+//       artists: track.artists.map((artist) => artist.name),
+//       popularity: track.popularity,
+//       externalUrl: track.external_urls.spotify,
+//       images: track.album.images,
+//     }));
 
-    // Response
-    return res.json({
-      message: "Success",
-      total_tracks: top10Tracks.length,
-      data: top10Tracks,
-    });
-  } catch (error) {
-    console.error("Error getting top tracks:", JSON.stringify(error, null, 4));
-    return res.status(500).json({
-      message: "Error getting top tracks",
-      error: error.response ? error.response.data : error.message,
-    });
-  }
-});
+//     // Response
+//     return res.json({
+//       message: "Success",
+//       total_tracks: top10Tracks.length,
+//       data: top10Tracks,
+//     });
+//   } catch (error) {
+//     console.error("Error getting top tracks:", JSON.stringify(error, null, 4));
+//     return res.status(500).json({
+//       message: "Error getting top tracks",
+//       error: error.response ? error.response.data : error.message,
+//     });
+//   }
+// });
 
 app.get("/detailed-stats", async (req, res) => {
   try {
